Fix minor code issues.

+++ src/preconditioner/ras_preconditioner.cpp
@@ -48,7 +48,7 @@ ras_preconditioner::ras_preconditioner(const model& model, const geometry::point
   }
 #if !POLATORY_RECOMPUTE_AND_CLEAR
 #pragma omp parallel for
-  for (size_t i = 0; i < fine_grids_.back().size(); i++) {
+  for (ptrdiff_t i = 0; i < fine_grids_.back().size(); i++) {
      auto& fine = fine_grids_.back()[i];
      fine.setup(points_);
   }
@@ -72,7 +72,7 @@ ras_preconditioner::ras_preconditioner(const model& model, const geometry::point
     }
 #if !POLATORY_RECOMPUTE_AND_CLEAR
 #pragma omp parallel for
-    for (size_t i = 0; i < fine_grids_.back().size(); i++) {
+    for (ptrdiff_t i = 0; i < fine_grids_.back().size(); i++) {
        auto& fine = fine_grids_.back()[i];
        fine.setup(points_);
     }
@@ -132,7 +132,7 @@ common::valuesd ras_preconditioner::operator()(const common::valuesd& v) const {
 #pragma omp parallel for schedule(guided)
       // Do not declare i to be of type size_t.
       //   https://docs.microsoft.com/cpp/error-messages/compiler-errors-2/compiler-error-c3016
-      for (int i = 0; i < fine_grids_[level].size(); i++) {
+      for (ptrdiff_t i = 0; i < fine_grids_[level].size(); i++) {
         auto& fine = fine_grids_[level][i];
 #if POLATORY_RECOMPUTE_AND_CLEAR
         fine.setup(points_);
